+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed a issue on iOS where the player crashed when it was destroyed while in fullscreen.
 
+### Changed
+
+- Moved the `hlsDateRange` property from `SourceConfiguration` to its correct location in `TypedSource`.
+
 ## [10.1.0] - 25-10-06
 
 ### Added