+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
-- Fixed a issue on iOS where the player crashed when it was destroyed while in fullscreen.
+- Fixed an issue on iOS where the player crashed when it was destroyed while in fullscreen.
 - Fixed an issue on Android where creating a caching task with `MediaCache.createTask` would not return the created `task`.
 
 ### Changed