+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed an issue on iOS where the player crashed when it was destroyed while in fullscreen.
 - Fixed an issue on Android where creating a caching task with `MediaCache.createTask` would not return the created `task`.
+- Fixed an issue where setting `player.theoLive.authToken = undefined` would cause a crash on Android.
 
 ### Changed